sprint6/desafio/requirements.txt
python-dotenv
boto3
datetime